Move patient upload code out of QBP client
