+++ test/iis_shc_issuer/qbp_client_test.rb
@@ -68,10 +68,4 @@ class QBPClientTest < ActiveSupport::TestCase
       end
     end
   end
-
-  # # WARNING: Running tests with this test uncommented could change sandbox data and cause other tests to fail
-  # # Temporary Test to upload a patient
-  # test 'Uploading a patient' do
-  #   IisShcIssuer::QBPClient.upload_patient() #Enter VXU Upload fixture path parameter here
-  # end
 end
