+++ test/iis_shc_issuer/vxu_client_test.rb
@@ -10,4 +10,9 @@ VCR.configure do |config|
 end
 
 class VXUClientTest < ActiveSupport::TestCase
+  # # WARNING: Running tests with this test uncommented could change sandbox data and cause other tests to fail
+  # # Temporary Test to upload a patient
+  # test 'Uploading a patient' do
+  #   IisShcIssuer::QBPClient.upload_patient() #Enter VXU Upload fixture path parameter here
+  # end
 end
